+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 See README.md before updating this file.
 
 ## Unreleased [#](https://git.cashnetusa.com/trogdor/landable/compare/v1.2.4...master)
+* Ability to add other non-Landable pages to the sitemap [#81]
 * current_page returns correct page for previews and published pages [#79]
 * URI.encode request.referer url [#80]